SolarSystem: Pass size as doubles
git-svn-id: https://codex.cs.bham.ac.uk/svn/team-project/D2@399 cfd7d107-c69a-4b07-9ecf-543bb1cfc744

+++ module-planets/src/d2/teamproject/module/planets/PlanetView.java
@@ -69,7 +69,7 @@ public class PlanetView extends VisualisationView {
 
     @Override
     public void onOpen() {
-        sSystem = new SolarSystem(controller.getPlanets(), (int) PARTH.WIDTH, (int) (PARTH.HEIGHT * 0.75), skybox);
+        sSystem = new SolarSystem(controller.getPlanets(), PARTH.WIDTH, PARTH.HEIGHT * 0.75, skybox);
         contentBox.getChildren().add(sSystem.getScene());
     }
